wip GetAuthValCode
GetAuthValCode

wip

replace ->count with != 0

from the doc:
If a resultset is used in a numeric context it returns the "count".
However, if it is used in a boolean context it is always true. So if you
want to check if a resultset has any results, you must use if $rs != 0.

fix qa issues

followup wip (...)

to squash