Introduce Koha::Object::Simple
This may be a crazy patch, but needed to share it.

I was trying to remove C4::AuthorisedValues and use
Koha::AuthorisedValues instead.
But C4::AuthorisedValues has a very good caching mechanism and
performances are excellent.

Removing it and replacing the occurrences to
Koha::AuthorisedValues->search will have terrible impacts on
performance, so we need to optimise the caching of our Koha::Objects.

I first took a look at what DBIx::Class provides: We can specify a
cursor_class which will use caching. But I did not get very good
performance boost, even if that can be useful in other cases.

So I tried to think about a mechanism based on Koha::Cache to cache the
whole AV list. I ended up with Koha::Object::Simple, which behave as a
Koha::Object, but is just a blessed hashref.
However the AUTOLOAD allows calls to Koha::AuthorisedValue method (but
can be dangerous if it is not controled).
The idea is to
  1. retrieve the AVs using DBIx::Class
  2. unblessed them to allow quick caching
  3. return an array of Koha::Object::Simple::AuthorisedValue
This is done on first call, then, if the things are cached:
  4. retrieve from cache
  5. generate the Koha::Object::Simple::AuthorisedValue list

At this point we have a problem: most of the time ->opac_description is
called on the AV. That will make generate a Koha::AuthorisedValue object
(from AUTOLOAD of Koha::Object::Simple). Which is bad for perf.
So, before caching, we call the method and store it. That way it is in
the hashref we set to/get from the cache.

Try `perl t/db_dependent/Koha/Object/Simple.t`